SandboxTestRunner is no longer concerned with reporting perf stats, that happens in RobolectricTestRunner.
